disable query caching and drop legacy comment
Responses generated by Lua records aren't cached so there's no longer a
load balancing issue with the packet cache, but it also makes the caches
less valuable. GeoIP with EDNS Client Subset enabled also substantially
reduces the usefulness of the caches and causes them to expand to a huge
size. Caching randomized dnscheck queries is wasteful and we found a way
to exclude them via a GeoIP service record but they're becoming a larger
and larger portion of the overall queries so there's little point in any
local caching in the long term. In the long term, we could split it out
into a separate higher performance DNS server with fewer capabilities if
it ever becomes a problem but it's currently far from being an issue.