use nameserver location as fallback for GeoDNS
This fills out missing GeoIP information based on the location of our
nameservers. We have Rage4 anycast with 4 locations for ns1 and BuyVM
anycast with 3 locations for ns2 so resolvers tend to be routed to our
nearest nameserver, giving us an approximation of the resolver location
which is generally near the client. This will almost always be better
than the previous approach of falling back to US East.