move GeoIP debugging to a special non-DNSSEC zone
This eliminates the last way to generate more than a small fixed amount
of DNSSEC signatures. It also avoids a hard dependency on online signing
since PowerDNS could gain support for offline signing with Lua records
only used to return a fixed set of records.

Moving this to a dedicated zone also makes it easier to make changes to
this for testing purposes without potentially breaking the rest of the
crucial grapheneos.org zone.S

This also moves it out from grapheneos.org to a placeholder domain more
suitable for debugging.