replace nginx with dnsdist for DNS-over-TLS
dnsdist provides support for passing along the resolver IP to PowerDNS
via PROXY protocol version 2 to avoid losing it as a fallback for EDNS
Client Subnet. This wasn't possible with nginx since it still only has
support for PROXY protocol version 1 with upstream servers even though
it supports listening with version 2.

We'll also be able to enable DNS-over-QUIC when we want to deploy it.