enable DNSSEC for dnscheck wildcard subdomain
This no longer requires an unbounded number of signatures since we
switched to the BIND backend with regular range-based NSEC support.

RFC 8198 allows resolvers to cache the positive answers for wildcards
based on DNSSEC wildcard support. This will greatly reduce the queries
for the dnscheck subdomain from resolvers supporting this part of RFC
8198.