ns1: purge Bharti everywhere but Mumbai and Singapore
Bharti sometimes sends to peer routes in far away locations rather than
a nearby customer route. Purging it from Singapore the routing in some
cases but results in some traffic being sent to Frankfurt to hand over
to Hurricane Electric. It would be nice if they preferred the customer
routes so we wouldn't need to create a partial fix.