remove Tempest London update server
The new management of Tempest has withdrawn the sponsorship for our
update server and we're going to switch providers.

Lacking an update server in Europe is a short term situation so we'll
use the Toronto server as a Europe location rather than updating our
GeoDNS setup to send more of Asia to Los Angeles.